# docs/swagger.yaml
basePath: /
definitions:
...
  model.Course:
    properties:
      codeCourse:
        type: string
      id:
        type: integer
      lessons:
        type: integer
      name:
        type: string
      subject:
        $ref: '#/definitions/model.Subject'
      subjectId:
        type: integer
      teacher:
        $ref: '#/definitions/model.Teacher'
      teacherId:
        type: integer
    type: object
...
  model.File:
    properties:
      course:
        $ref: '#/definitions/model.Course'
      courseId:
        type: integer
      id:
        type: integer
      link:
        type: string
      name:
        type: string
    type: object
...
  model.Subject:
    properties:
      codeSubject:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  model.Teacher:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      pass:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: Api canvas project
  title: Canvas Project
  version: "2.0"
paths:
...
  /teacher/courses/{id}/file:
    get:
      consumes:
      - application/json
      description: Teacher get all file of course
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with file
  /teacher/courses/{id}/file/{file_id}:
    get:
      consumes:
      - application/json
      description: teacher get a file of course
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      - description: file id
        in: path
        name: file_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with file
  /teacher/courses/{id}/file/create:
    post:
      consumes:
      - application/json
      description: Teacher create file for course
      parameters:
      - description: Course id
        in: path
        name: id
        required: true
        type: string
      - description: File
        format: model.File
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.File'
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with file
  /teacher/courses/{id}/file/delete:
    delete:
      consumes:
      - application/json
      description: Teacher delete file of course
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      - description: File need delete
        format: model.File
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.File'
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with file
  /teacher/courses/{id}/file/update:
    put:
      consumes:
      - application/json
      description: Teacher update file of course
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      - description: File need update
        format: model.File
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.File'
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with file
swagger: "2.0"